SSCS-7483 Refactor for email service
