+++ src/main/java/uk/gov/hmcts/reform/sscs/service/EmailService.java
@@ -10,7 +10,6 @@ import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.retry.annotation.Backoff;
 import org.springframework.retry.annotation.Retryable;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Appellant;
 import uk.gov.hmcts.reform.sscs.domain.email.Email;
 import uk.gov.hmcts.reform.sscs.domain.email.EmailAttachment;
 import uk.gov.hmcts.reform.sscs.exception.EmailSendFailedException;
@@ -18,8 +17,6 @@ import uk.gov.hmcts.reform.sscs.exception.EmailSendFailedException;
 @Service
 @Slf4j
 public class EmailService {
-
-    private static final String ID_FORMAT = "%s_%s";
 
     private final JavaMailSender javaMailSender;
 
@@ -63,11 +60,5 @@ public class EmailService {
             log.error("Error while sending email {} ", e);
             throw new EmailSendFailedException("Error while sending email", e);
         }
-    }
-
-    public String generateUniqueEmailId(Appellant appellant) {
-        String appellantLastName = appellant.getName().getLastName();
-        String nino = appellant.getIdentity().getNino();
-        return String.format(ID_FORMAT, appellantLastName, nino.substring(nino.length() - 3));
     }
 }
